Ui changes, Username Class, CryptManager, databasequerier
Ui chanages consist of chanages in the new application form. The combo box now takes a list of root nodes from the password manager form and populates itself. The program on start up will check for a secret key (TODO) then ask if user wants to point to a key they already have or auto-generate one.

The username class is now a thing. Takes in a username and password(encrypted).

Crypt manager methods are now static as well as the class was moved to the shared project.

The database Querier has a getApptypes method which returns the list of distinct apptypes for a specific user.