NetworkingProject  and Password Manager
Initial push for the networking project. It contains a base message class along with a message sender class. The base message class is intended to be a parent class for all messages to inherit from. The messageheader class contains an id and a MAGICNUMBER memebers.
Fixed the TODO in the Password manager form. It now checks for keyFile to see if present. Also the question is phrased properly.