SocketManager class
the SocketManager class is a singleton on the clientside used to get the socket and establish a connection on start up. This will be used anytime the MessageSender class needs to send a message.

Ui changes were added to the logindialog to handle the case where you try to connect to a server and need a host and port.