Declare MemoryStream in using block
It is best practice to declare any variables that implements IDisposable
(e.g., MemoryStream) inside a using block so that IDisposable.Dispose()
is called on the class before the scope exits (of course, there are
exceptions to this rule).

At any rate, the MemoryStream should be closed as soon as
MessageHeader.toByteArray() is done (since we won't be using the stream
once we get the resulting byte[]), so declare the MemoryStream inside a
using block.