Server.cs ClientSession ApplicationsResponse
A Server class inside the passwrod manager server project was introduced. It has two threads one listening for connections and one handling messages from current connctions. A class called client session was created.
This class contains the username of the client that is logged in. This is used in pretty much all of the queries in the database querier.
The Applications response is a response containing the list of applications for a user.