Introduce a MessageUtils class
This utility class contain functions for serializing a MessageBase into
a byte[]. This byte[] containso not only the message but also the proper
header needed for the Password Manager to know that this is a message
that was sent from the client or the server.

In addition, remove the MessageSender class (since this class pretty
much takes over).