Changed MessageHandler & ReadyClient class
The big changes were when the message handler handle message method gets called it now takes in a socket. The pendingTasks list is now a list of ReadyCLients which contains a task and a socket so that we know which sockets go with which tasks and where to send the message to!