Ensure there is enough data to read in header + payload before handling
Prevent some possible issues that may crop up when we attempt to send a
long message through the TCP socket, but TCP breaks up our message,
causing the server to not get the full message immediately.

Note that there is still some gotchas that needs to be kept in mind
(i.e., keeping track of the header that was last read in for sockets
that are currently sending a huge payload message such that it'll
require multiple reads to get the entire message, properly storing a
large payload in the buffer without overwriting some existing data,
etc.).