Hide LoginDialog and show PasswordManagerForm upon getting applications
Note that we have to invoke LoginDialog.Hide and
PasswordManagerForm.Show inside a lambda function that'll be called by
Form.Invoke. This is because GUI-related operations (like showing or
hiding forms) must be done on the GUI thread (which will occur if we
wrap those call inside a lambda function that'll be called by
Form.Invoke).

If we don't, it causes the form to freak out and errors to occur, since
GUI operations must occur in the same thread that the GUI was created.