Use e.Node.Level to determine if a username node is selected
Don't use e.Node.Nodes.Count, which can be misleading if the user so
happens to delete all their username, and suddenly the application node
has been selected.

(In addition, rename isAppSelected to isUserSelected. The previous name
was a bit misleading, since it's the username node we're looking for,
not the app node).