Creating new users
The request and response for creating a new user has been implented. Also the client session has been modified to now hold a boolean indicating rather or not the user with the session is an admin. The database querier has been updated with the corresponding queries to add a user as well as checking if a user is admin