Use the new SocketManager.SendRequest<TResponse> method in LoginDialog
In addition, update the "Invalid username/password" message to use the
TaskDialog API.

This commit should also fix an issue where the "Invalid
username/password" message was showing even if the user had already
logged in when the server forces a disconnect. This is due to the
LoginDialog not removing the event handler from
SocketManager.Disconnect.

Now that we're using async/await to get the LoginResponse and determine
if the username or password is incorrect (since
SocketManager.SendRequest<LoginResponse> will return null to indicate
the server forced the client to disconnect), we don't need to worry
about removing our event handler properly.