Remove the Task for reading messages
This was interfering with the Socket.ReceiveAsync method that's used in
SocketManager.SendRequest (since either the Socket.Poll that was inside
the background Task or the Socket.RecevieAsync would be called when a
message was received from the server).

Since SocketManager.SendRequest will eventually supersede
SocketManager.SendMessage, remove the background Task that's responsible
for handling responses for requests sent with SocketManager.SendMessage.

Note that this means that SocketManager.SendMessage is broken and the
MessageHandler is broken and not in use (for now).