Prefer SocketManager.SendRequest<T> instead of SendResponse
This way, we can use async/await to await for the response that we'll
receive after sending the request (via SocketManager.SendRequest) and
handle the response immediately after the await.

This also make the code easier to follow (in that one does not have to
hop from the PasswordManagerForm to see how the request is sent and then
go to the MessageHandler to see how the response is handled/what happens
afterward).