edit app request and change to change app type request
There has been a change to the ChangeAppTypeRequest so that now it can take in an array of applications.
The message handler on the server has also been made to handle this new change. This is so the ui can
edit an apptype for several applications at once in the tree node.
Also EditApplicationRequest has been added and the databaseQuerier and MessageHandler classes have been
changed to handle it