Move "Generate key file" dialog confirmation into LoginDialog
Instead of asking the user if they want to generate a key file within
the LoginDialog.FormClosed delegate in Program.ShowLoginDialog, handle
this within the LoginDialog itself.

That way, we keep code related to the LoginDialog (i.e., generating a
key file) in one place, instead of leaking it over to the
Program.ShowLoginDialog method. In addition, this fix the rather awkward
transition of the LoginDialog disappearing and reappearing if the user
decides to not generate a key file (which causes the LoginDialog to show
up again).

In addition, fix an exception that was thrown whenever the user
successfully logs in, but didn't generate a key file, and attempts to
press Login again, causing the client to throw an exception stating that
the user attempted to connect to an already-connected socket.