Add ON DELETE CASCADE to applications table
Make sure that when a user is deleted from the users table that their
applications (inside hte applications table) are also deleted.

In addition, be sure to run PRAGMA foreign_keys = ON so that foreign
keys constraints are applied (because in SQLite3, foreign key
constraints aren't enabled by default for some reason).