Preset the app type based off the selected node when adding an app
When the user presses "Add Application", determine what application type
has been selected (or if an application or username has been selected,
determine the application type of the selected application) and set
the application type to the selected type in the NewApplicationForm.

That way, if a user has selected an app that belongs to an app type
(e.g., Facebook, which is under Social Media), then the
NewApplicationForm will automatically select Social Media when they
press "Add Application" (operating under the assumption that the user
wants to add a new application or username that'll be under the Social
Media type).