Stop the client from crashing when a bad login occurs
Subscribe to the SocketManager.Disconnected event (so that the GUI
can show an error message if the server forces the client to
disconnect suddenly) when the PasswordManagerForm is shown for the first
time.

Don't do it before showing the LoginDialog. Otherwise, the above
error message will be shown (instead of the "Bad username/login"
message), and the program will suddenly quit (instead of allowing
the user to try authenticating again).